+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 * `RedundantSelf` now allows `self.` followed by any ruby keyword
 * [#391](https://github.com/bbatsov/rubocop/issues/391) - Fix bug in counting slashes in a regexp.
 * [#394](https://github.com/bbatsov/rubocop/issues/394) - `DotPosition` cop handles correctly code like `l.(1)`
+* [#390](https://github.com/bbatsov/rubocop/issues/390) - `CommentAnnotation` cop allows keywords (e.g. Review, Optimize) if they just begin a sentence.
 
 ## 0.10.0 (17/07/2013)